refactor: move hash and path functions to utils
